Add value as new first arg of Document constructor
BREAKING CHANGE: `options` is now the second rather than first argument
for `new YAML.Document(value, options)`. If you previously had:
    const doc = new YAML.Document({ version: '1.1' })
    doc.setSchema()
    doc.contents = doc.schema.createNode({ some: 'value' })
you should replace that with:
    const doc = new YAML.Document({ some: 'value' }, { version: '1.1' })

Calling the constructor with no arguments has the exact same behaviour
as previously. Calling it with a defined value will initialise its
`contents` with that value, as well as initialising the `schema`.