When creating a mapping from a JS Object, drop undefined values (Closes #173)
BREAKING CHANGE: When creating a YAML mapping from a JS Object, keys
with `undefined` values are not included. This matches the behaviour of
`JSON.stringify()`. In all other cases, the handling of `undefined` is
unchanged, so e.g. `[1, undefined, 2]` will still result in a sequence
of three items, the second of which has a `null` value. This too matches
the behaviour of `JSON.stringify()`.

Previously, `undefined` object values were mapped to `null`.