Drop the keepCstNodes and keepNodeTypes options
BREAKING CHANGE: The earlier parser change has meant that information is
retained differently during composition. Effectively, the values of
these options are no longer used, so they should be explicitly dropped
from the options interfaces to make that clear. The behaviour now is as
if `{ keepCstNodes: false, keepNodeTypes: true }` is always set.

This may require some API changes for CST users, as errors also no
longer provide a reference to the source CST node, just the original
line position. To match an error to a node, you should search for a node
that includes the error position in its range.

To discard node type information from the tree, use a visitor:

    import { isNode, parseDocument, visit } from 'yaml'

    const doc = parseDocument('{ foo: "bar" }')
    visit(doc, (key, node) => {
      if (isNode(node)) delete node.type
    })

    String(doc) === 'foo: bar\n'