Refactor doc.setSchema(); drop Document.defaults
BREAKING CHANGE: The new API of doc.setSchema(), now requiring an
explicit version as well as supporting all Schema options. This is
required to reduce the dependency on a mutating doc.options object.

-doc.setSchema('json')
+doc.setSchema('1.2', { schema: 'json' })

-doc.options.merge = false
-doc.setSchema('1.1')
+doc.setSchema('1.1', { merge: false })

As a side effect of these changes, the Document.defaults object is
removed, along with the doc.getDefaults() method. The version-specific
schema options now need to be set explicitly in function arguments.