feat!: Make Pair not extend NodeBase; drop its prop forwarding
BREAKING CHANGE: While the inheritance of the class is just an
implementation detail, dropping the prop forwarding means that a user
that previously set the `commentBefore`, `comment` or `spaceBefore`
property of a pair will now need to set it for the appropriate key or
value.

While adjusting the test suite for the changes, the somewhat surprising
behaviour of multi-line comments on scalar values getting handled as
`commentBefore` instead of `comment` is also dropped.